mem noAction&low&high

+++ FileHelper.h
@@ -15,7 +15,7 @@ public:
     template <typename T>
     static void write(const std::string &file_name, T content, bool overwrite);
 
-    static void readProcMemInfo(vector<int> &result);
+    static void readProcMemInfo(vector<long> &result);
 };
 
 #endif //ERRORGENERATOR_FILEOPERATION_H
